% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_descriptions.R
\docType{data}
\name{campaign_descriptions}
\alias{campaign_descriptions}
\title{Campaign metadata.}
\format{A data frame with 30 rows and 4 variables
\describe{
\item{description}{Type of campaign (TypeA, TypeB, TypeC)}
\item{campaign}{Uniquely identifies each campaign; Ranges 1-30}
\item{start_day}{Start date of campaign}
\item{end_day}{End date of campaign}
}}
\source{
84.51, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
campaign_descriptions
}
\description{
Campaign metadata for all campaigns run for the Customer Journey study. This
dataset gives the length of time for which a campaign runs. So, any coupons
received as part of a campaign are valid within the dates contained in this
dataset.
}
\examples{
\dontrun{
if (require("dplyr")) {
campaign_descriptions

# Join product campaign metadata to campaign_table dataset
campaigns \%>\%
  left_join(campaign_descriptions, "campaign")
}
}
}
\keyword{datasets}
