% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_table.R
\docType{data}
\name{campaign_table}
\alias{campaign_table}
\title{Campaign to household data.}
\format{A data frame with 7,208 rows and 3 variables
\describe{
\item{description}{Type of campaign (TypeA, TypeB, TypeC)}
\item{household_key}{Uniquely identifies each household}
\item{campaign}{Uniquely identifies each campaign; Ranges 1-30}
}}
\source{
84.51, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
campaign_table
}
\description{
Data on the campaigns received by each household in the Complete Journey study.
Each household received a different set of campaigns.
}
\examples{
\dontrun{
if (require("dplyr")) {
campaign_table

# Join household metadata to campaign_table dataset
campaign_table \%>\%
  left_join(hh_demographics, "household_key")
}
}
}
\keyword{datasets}
