% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completejourney.R
\docType{package}
\name{completejourney}
\alias{completejourney}
\alias{completejourney-package}
\title{\code{completejourney} package}
\description{
Retail shopping transactions for 2,469 households over one year
}
\details{
Learn more here:
\href{https://github.com/bradleyboehmke/completejourney}{GitHub}
}
