% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupon_redemption.R
\docType{data}
\name{coupon_redemption}
\alias{coupon_redemption}
\title{Coupon redemption data.}
\format{A data frame with 2,318 rows and 4 variables
\describe{
  \item{household_key}{Uniquely identifies each household}
  \item{day}{Day when transaction occurred}
  \item{coupon_upc}{Uniquely identifies each coupon (unique to household and campaign)}
  \item{campaign}{Uniquely identifies each campaign}
}}
\source{
84.51, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
coupon_redemption
}
\description{
Coupon data identifying the coupons that each household redeemed in the Complete
Journey study.
}
\examples{
\dontrun{
if (require("dplyr")) {
coupon_redempt

# Join coupon metadata to coupon_redempt dataset
coupon_redemption \%>\%
  left_join(coupons, "coupon_upc")
}
}
}
\keyword{datasets}
