% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupon_redemptions.R
\docType{data}
\name{coupon_redemptions}
\alias{coupon_redemptions}
\title{Coupon redemption data.}
\format{A data frame with 2,318 rows and 4 variables
\describe{
  \item{household_id}{Uniquely identifies each household}
  \item{coupon_upc}{Uniquely identifies each coupon (unique to household and campaign)}
  \item{campaign_id}{Uniquely identifies each campaign}
  \item{redemption_date}{Day when transaction occurred}
}}
\source{
84.51, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
coupon_redemptions
}
\description{
Coupon data identifying the coupons that each household redeemed in the Complete
Journey study.
}
\examples{
\dontrun{
if (require("dplyr")) {
coupon_redemptions

# Join coupon metadata to coupon_redempt dataset
coupon_redemptions \%>\%
  left_join(coupons, "coupon_upc")
}
}
}
\keyword{datasets}
