% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\docType{data}
\name{demographics}
\alias{demographics}
\title{Household demographic metadata.}
\format{A data frame with 801 rows and 8 variables
\describe{
\item{household_id}{Uniquely identifies each household}
\item{age_desc}{Estimated age range}
\item{marital_status_code}{Marital status (A - Married, B - Single, U - Unknown)}
\item{income_desc}{Household income}
\item{homeowner_desc}{Homeowner status (Homeowner, Renter, Unknown)}
\item{hh_comp_desc}{Household composition description}
\item{household_size_desc}{Size of household up to 5+}
\item{kid_category_desc}{Number of children present up to 3+}
}}
\source{
84.51, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
demographics
}
\description{
Household demographic metadata for households participating in the Customer Journey study.
Due to nature of the data, the demographic information is not available for all
households.
}
\examples{
\dontrun{
if (require("dplyr")) {
demographics

# Transaction line items that don't have household metadata
transactions \%>\%
  anti_join(demographics, "household_id")
}
}

}
\keyword{datasets}
