% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download Complete Journey Data}
\source{
Downloading from \url{https://github.com/bradleyboehmke/completejourney/tree/master/data}.
  Data originated from 84.51°, Customer Journey study, \url{http://www.8451.com/area51/} 
  and were processes for analysis.
}
\usage{
get_data(which = "all", verbose = TRUE)
}
\arguments{
\item{which}{Character string of one or more data sets to be downloaded.
Can be one of the following; default is \code{"all"}:
\itemize{
    \item{"all"}
    \item{"compaigns"}
    \item{"campaign_descriptions"}
    \item{"coupons"}
    \item{"coupon_redemptions"}
    \item{"demographics"}
    \item{"products"}
    \item{"promotions"}
    \item{"transactions"}
   }}

\item{verbose}{Logical indicator whether or not to download silently.}
}
\value{
Each downloaded data set is saved as a tibble in the users global 
  environment. For specifc details on a given data set see the data sets 
  respective help file (i.e. \code{?transactions}).
}
\description{
The data sets available through this package are quite sizeable; and too 
large to be contained within the package. \code{get_data()} provides an
efficient method for downloading one or more of the data sets from the
source GitHub repository.
}
\examples{
\donttest{
# download all data sets
get_data(which = "all")

# download transactions data
get_data(which = "transactions")

# download multiple data sets
get_data(which = c("transactions", "promotions"))
}
}
