% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%<-\%}
\alias{\%<-\%}
\title{Assign values to names}
\usage{
x \%<-\% value
}
\arguments{
\item{x}{A name structure.}

\item{value}{A list of values, vector of values, or R objects to assign.}
}
\description{
See \code{\link[zeallot]{\%<-\%}} for more details.
}
