% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coupons.R
\docType{data}
\name{coupons}
\alias{coupons}
\title{Coupon metadata.}
\format{
A data frame with 116,204 rows and 3 variables
\itemize{
\item coupon_upc: Uniquely identifies each coupon (unique to household and campaign)
\item product_id: Uniquely identifies each product
\item campaign_id: Uniquely identifies each campaign
}
}
\source{
84.51°, Customer Journey study, \url{http://www.8451.com/area51/}
}
\usage{
coupons
}
\value{
\item{coupons}{a tibble}
}
\description{
Coupon metadata for all coupons used in campaigns advertised to households
participating in the Customer Journey study.
}
\examples{
\donttest{
# full data set
coupons

# Join product metadata to coupon dataset
require("dplyr")
coupons \%>\%
  left_join(products, "product_id")
}
}
\keyword{datasets}
