% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Download Complete Journey Data}
\source{
Downloading from \url{https://github.com/bradleyboehmke/completejourney/tree/master/data}.
  Data originated from 84.51°, Customer Journey study, \url{http://www.8451.com/area51/} 
  and were processes for analysis.
}
\usage{
get_data(which = "all", verbose = TRUE)
}
\arguments{
\item{which}{Character string of one or more data sets to be downloaded.
Can be one of the following; default is \code{"all"}:
\itemize{
    \item{"all"}
    \item{"compaigns"}
    \item{"campaign_descriptions"}
    \item{"coupons"}
    \item{"coupon_redemptions"}
    \item{"demographics"}
    \item{"products"}
    \item{"promotions"}
    \item{"transactions"}
   }}

\item{verbose}{Logical indicator whether or not to download silently.}
}
\value{
Downloading a single data set will result in a tibble whereas 
  downloading multiple data sets will return a list containing each tibble. 
  For specific details on a given data set see the data sets respective help 
  file (i.e. \code{?transactions}).
}
\description{
The data sets available through this package are quite sizeable; and too 
large to be contained within the package. \code{get_data()} provides an
efficient method for downloading one or more of the data sets from the
source GitHub repository.
}
\examples{
\donttest{
# download transactions data
transactions <- get_data(which = "transactions", verbose = FALSE)

# download multiple data sets
c(transactions, promotions) \%<-\% get_data(which = c("transactions", "promotions"))

# download all data sets?
c(campaigns, campaign_descriptions, coupons,
  coupon_redemptions, demographics, products,
  promotions, transactions) \%<-\% get_data(which = "all")

}
}
\seealso{
Use \code{\link[zeallot]{\%<-\%}} for unpacking a list with multiple 
tibbles to their own global environment tibble.
}
